Wrote Backend Logic to the Selenium Program
Finished writing the back-end code for selenium. The code runs all the way up til basically the end of the event request where manual inputs are basically necessary , but it gets through all of the tedious tasks amazingly. Added a couple TODO's and will add those in as issues as well.